Refactored directory structure a bit more
