+++ rascal-vscode-extension/src/fs/RascalFileSystemProviders.ts
