+++ rascal-vscode-extension/src/fs/VSCodeURIResolver.ts
